feat(charserv): function converts str to int before storing

+++ src/services/character_service.py
@@ -90,7 +90,7 @@ class CharacterService:
         Args:
             character_id(int): Muokattavan hahmon tunnisteluku.
             statistic(str): Ominaisuus, jota halutaan muokata.
-            new_value(float): Uusi arvo muokattavalle ominaisuudelle.
+            new_value(int): Uusi arvo muokattavalle ominaisuudelle.
         """
 
         if not new_value.isdigit():
@@ -100,7 +100,7 @@ class CharacterService:
             raise NegativeValueError("Value must be positive")
 
         self._character_repository.update_character_property(
-            character_id, statistic, new_value)
+            character_id, statistic, int(new_value))
 
 
 character_service = CharacterService()
